requirements.txt
Django==4.2.20
Pillow==10.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-import-export==3.3.5
whitenoise==6.9.0
gunicorn==21.2.0
requests==2.31.0
python-decouple==3.8
django-crispy-forms==2.1.0
